+++ src/tbi_markov/__init__.py
@@ -3,4 +3,4 @@
 from .hmm import DiagonalGaussianHMM
 
 __all__ = ["DiagonalGaussianHMM"]
-__version__ = "1.1.0"
+__version__ = "1.2.0"